Move the actions directory and update the paths
