+++ src/components/Register_Login_forms/Login.jsx
@@ -1,7 +1,7 @@
 import { useForm } from 'react-hook-form';
 import { useDispatch } from 'react-redux';
 import { useNavigate } from 'react-router-dom';
-import sendLogin from '../../redux/actions/sendLogin';
+import sendLogin from '../../redux/user/actions/sendLogin';
 import EmailInput from './inputs/EmailInput';
 import PasswordInput from './inputs/PasswordInput';
