+++ src/redux/user/actions/sendLogout.js
