+++ src/redux/user/actions/sendRegistration.js
