+++ src/redux/user/userSlice.js
@@ -1,11 +1,11 @@
 import { createSlice } from '@reduxjs/toolkit';
-import sendRegistration from '../actions/sendRegistration';
-import sendLogin from '../actions/sendLogin';
-import sendLogout from '../actions/sendLogout';
+import sendRegistration from './actions/sendRegistration';
+import sendLogin from './actions/sendLogin';
+import sendLogout from './actions/sendLogout';
 
 const userSlice = createSlice({
   name: 'user',
-  initialState: { 
+  initialState: {
     data: null,
     error: null,
     notice: null,
@@ -43,8 +43,8 @@ const userSlice = createSlice({
       .addCase(sendLogout.rejected, (state, action) => {
         state.notice = null;
         state.error = action.payload.message;
-      })
-  },
+      });
+  }
 });
 
 export default userSlice.reducer;
